==> scripts/bronze_to_silver/1_b2s_column_filtering.ipynb <==
import pandas as pd
from pathlib import Path

# Get project root (go up from scripts/bronze_to_silver to project root)
PROJECT_ROOT = Path(__file__).parent.parent.parent if '__file__' in globals() else Path.cwd().parent.parent
file_path = PROJECT_ROOT / "data" / "bronze" / "bronze_dataset.csv"

df = pd.read_csv(file_path)
df.head()

df.columns

cols_to_keep = ['subject_id',
 'hadm_id',
  'admittime',
  'dischtime',
 'gender',
 'anchor_age',
 'admission_type',
 'admission_location',
 'discharge_location',
 'race',
 'catheter_present',
 'final_insertion_date',
 'final_removal_date',
 'catheter_duration_days',
 'BMI',
 'diabetes',
 'cancer',
 'chronic_kidney_disease',
 'neurogenic_bladder',
 'recurrent_uti',
 'spinal_cord_injury',
 'neurological_disorder',
 'benign_prostatic_hyperplasia',
 'charlson_score',
 'length_of_stay',
 'num_of_transfers',
 'surgical_admission',
 'recent_urologic_abdominal_surgery',
 'icu_admission',
 'mobility_status',
 'catheter_type',
 'catheter_size_from_notes',
 'catheter_size',
 'urinary_obstruction_present',
 'improper_drainage_position',
 'n_catheter_manip_unique_types',
 'catheter_insertion',
 'catheter_removal',
 'catheter_removal_replacement',
 'no_of_invasive_devices',
 'multiple_invasive_devices',
 'catheter_indication',
 'catheter_care',
 'closed_system',
 'securement_device',
 'antibiotics_per_admission',
 'recent_antibiotic_use',
 'steroids_per_admission',
 'immunosuppressives_per_admission',
 'ppi_per_admission',
 'urinalysis_wbc',
 'urinalysis_rbc',
 'urinalysis_nitrite',
 'urine_bacteria',
 'blood_wbc',
 'creatinine',
 'blood_crp',
 'procalcitonin_measured',
 'urine_output_ml',
 'oliguria',
 'urine_culture_performed',
 'cfu_count',
 'gram_negative_organisms_present',
 'gram_positive_organisms_present',
 'fungi_present',
 'blood_culture_performed',
 'temperature',
 'heart_rate',
 'resp_rate',
 'o2sat',
 'BP_systolic',
 'BP_diastolic',
 'other_uti',
 'other_uti_present',
 'has_cauti_history',
 'pain_documented',
 'ventilator_used',
 'y']

df_filtered = df[cols_to_keep]

save_path = PROJECT_ROOT / "data" / "silver" / "bronze_columns_filtered.csv"
df_filtered.to_csv(save_path, index=False)
print("Filtered dataset saved at:", save_path)
