==> scripts/bronze_to_silver/6_normalize_features.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import MultiLabelBinarizer,StandardScaler

file_path = r"C:\Users\Coditas\Desktop\Projects\Cauti\silver\bronze_one_hot_encoded.csv"

df = pd.read_csv(file_path)
df.head()

df["catheter_size_fr"] = df["catheter_size_fr"].fillna(-1)

import numpy as np
from sklearn.preprocessing import StandardScaler

# Columns to exclude
exclude_cols = ['subject_id', 'hadm_id']

# Step 1: numeric columns (excluding ID columns)
numeric_cols = [
    col for col in df.select_dtypes(include=[np.number]).columns
    if col not in exclude_cols
]

# Step 2: automatically detect binary columns (only 0 and 1)
binary_cols = [
    col for col in numeric_cols
    if df[col].dropna().isin([0, 1]).all()
]

# Step 3: continuous = numeric but NOT binary
continuous_cols = list(set(numeric_cols) - set(binary_cols))

# Step 4: standardize continuous columns
scaler = StandardScaler()
df[continuous_cols] = scaler.fit_transform(df[continuous_cols])

print("Excluded columns:", exclude_cols)
print("Binary columns detected:", binary_cols)
print("Continuous columns detected:", continuous_cols)

save_path = r"C:\Users\Coditas\Desktop\Projects\Cauti\silver\silver_dataset.csv"
df.to_csv(save_path, index=False)

print("Filtered dataset saved at:", save_path)
